+++ types/src/main/java/com/cobber/fta/HllSketchDeserializer.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2017-2026 Tim Segall
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.cobber.fta;
+
+import java.io.IOException;
+import java.util.Base64;
+
+import org.apache.datasketches.hll.HllSketch;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+
+public class HllSketchDeserializer extends JsonDeserializer<HllSketch> {
+	@Override
+	public HllSketch deserialize(final JsonParser p, final DeserializationContext ctx) throws IOException {
+		return HllSketch.heapify(Base64.getDecoder().decode(p.getText()));
+	}
+}